Add constellation-boundary file to download script
Also, simplify and refactor the script itself; in particular, refuse to
overwrite existing files.

+++ download_data.py
@@ -1,18 +1,23 @@
-import os, requests
+import os, urllib2
 
 DATA_SETS = [
-('hygfull.csv', 'https://raw.github.com/astronexus/HYG-Database/master/hygfull.csv'),
-('ELEMENTS.NUMBR.gz', 'http://ssd.jpl.nasa.gov/dat/ELEMENTS.NUMBR.gz'),
-('ELEMENTS.UNNUM.gz', 'http://ssd.jpl.nasa.gov/dat/ELEMENTS.UNNUM.gz'),
-('ELEMENTS.COMET', 'http://ssd.jpl.nasa.gov/dat/ELEMENTS.COMET'),
-('visual.txt', 'http://www.celestrak.com/NORAD/elements/visual.txt')]
+    'ftp://cdsarc.u-strasbg.fr/pub/cats/VI/49/bound_18.dat',
+    'https://raw.github.com/astronexus/HYG-Database/master/hygfull.csv',
+    'http://ssd.jpl.nasa.gov/dat/ELEMENTS.NUMBR.gz',
+    'http://ssd.jpl.nasa.gov/dat/ELEMENTS.UNNUM.gz',
+    'http://ssd.jpl.nasa.gov/dat/ELEMENTS.COMET',
+    'http://www.celestrak.com/NORAD/elements/visual.txt',
+    ]
 
-def get_data_set(fname, url):
+def get_data_set(url):
+    fname = url.split('/')[-1]
+    if os.path.isfile(fname):
+        return
+    data = urllib2.urlopen(url).read()
     with open(fname, 'wb') as f:
-        r = requests.get(url)
-        f.write(r.content)
+        f.write(data)
 
 if __name__ == '__main__':
     os.chdir('data')
-    for data_name, location in DATA_SETS:
-        get_data_set(data_name, location)
+    for url in DATA_SETS:
+        get_data_set(url)
